Move objects into own modules
