+++ dds/src/dcps/dcps_domain_participant/builtin_publisher.rs
@@ -1,5 +1,6 @@
 use crate::{
     builtin_topics::{DCPS_PARTICIPANT, DCPS_PUBLICATION, DCPS_SUBSCRIPTION, DCPS_TOPIC},
+    dcps::dcps_domain_participant::builtin_data_writer::DataWriterEntity,
     infrastructure::{
         instance::InstanceHandle,
         qos::DataWriterQos,
@@ -18,7 +19,7 @@ use crate::{
 use alloc::string::String;
 
 use super::{
-    DataWriterEntity, ENTITYID_BUILTIN_PUBLISHER, ENTITYID_SEDP_BUILTIN_PUBLICATIONS_ANNOUNCER,
+    ENTITYID_BUILTIN_PUBLISHER, ENTITYID_SEDP_BUILTIN_PUBLICATIONS_ANNOUNCER,
     ENTITYID_SEDP_BUILTIN_SUBSCRIPTIONS_ANNOUNCER, ENTITYID_SEDP_BUILTIN_TOPICS_ANNOUNCER,
     ENTITYID_SPDP_BUILTIN_PARTICIPANT_WRITER, ENTITYID_TL_SVC_REPLY_WRITER,
     ENTITYID_TL_SVC_REQ_WRITER, TYPE_LOOKUP_REPLY_TOPIC_NAME, TYPE_LOOKUP_REQUEST_TOPIC_NAME,
