+++ dds/src/dcps/dcps_domain_participant/discovery_methods.rs
@@ -27,7 +27,9 @@ use crate::{
             ENTITYID_SEDP_BUILTIN_SUBSCRIPTIONS_DETECTOR, ENTITYID_SEDP_BUILTIN_TOPICS_ANNOUNCER,
             ENTITYID_SEDP_BUILTIN_TOPICS_DETECTOR, ENTITYID_TL_SVC_REPLY_READER,
             ENTITYID_TL_SVC_REPLY_WRITER, ENTITYID_TL_SVC_REQ_READER, ENTITYID_TL_SVC_REQ_WRITER,
-            IncompatibleSubscriptions, RtpsReader, UserDefinedDataReader, UserDefinedDataWriter,
+            RtpsReader, builtin_data_writer::IncompatibleSubscriptions,
+            user_defined_data_reader::UserDefinedDataReader,
+            user_defined_data_writer::UserDefinedDataWriter,
         },
         listeners::domain_participant_listener::ListenerMail,
     },
