+++ dds/src/dcps/dcps_domain_participant/publisher_methods.rs
@@ -2,7 +2,9 @@ use alloc::string::String;
 
 use crate::{
     dcps::{
-        dcps_domain_participant::{DcpsDomainParticipant, UserDefinedDataWriter},
+        dcps_domain_participant::{
+            DcpsDomainParticipant, user_defined_data_writer::UserDefinedDataWriter,
+        },
         listeners::{
             data_writer_listener::DcpsDataWriterListener, publisher_listener::DcpsPublisherListener,
         },
